Changing the saved polygon area from an int to a float.

+++ src/db.h
@@ -380,7 +380,7 @@ typedef struct _CADRow {
     int line_color;             // Border color
     int line_type;              // Border linetype
     int line_width;             // Border line width
-    int computed_area;          // Area in square kilometers
+    float computed_area;        // Area in square kilometers
     float raw_probability;      // Probability of area (POA) or probability of detection (POD)
     long label_latitude;        // Latitude for label placement
     long label_longitude;       // Longitude for label placement
